+++ R/RcppExports.R
@@ -45,6 +45,14 @@ rcpp_path_geojson <- function(data, params, geometry_columns, digits, layer_name
     .Call(`_mapdeck_rcpp_path_geojson`, data, params, geometry_columns, digits, layer_name)
 }
 
+rcpp_path_interleaved <- function(sf, params, list_columns, digits, layer_name) {
+    .Call(`_mapdeck_rcpp_path_interleaved`, sf, params, list_columns, digits, layer_name)
+}
+
+rcpp_trips_interleaved <- function(sf, params, list_columns, digits, layer_name, start_time) {
+    .Call(`_mapdeck_rcpp_trips_interleaved`, sf, params, list_columns, digits, layer_name, start_time)
+}
+
 rcpp_path_polyline <- function(data, params, geometry_columns, layer_name) {
     .Call(`_mapdeck_rcpp_path_polyline`, data, params, geometry_columns, layer_name)
 }
@@ -69,8 +77,8 @@ rcpp_polygon_geojson <- function(data, params, geometry_columns, digits) {
     .Call(`_mapdeck_rcpp_polygon_geojson`, data, params, geometry_columns, digits)
 }
 
-rcpp_triangle_columnar <- function(data, params, list_columns, geometry_columns, digits) {
-    .Call(`_mapdeck_rcpp_triangle_columnar`, data, params, list_columns, geometry_columns, digits)
+rcpp_triangle_interleaved <- function(data, params, list_columns, digits, layer_name) {
+    .Call(`_mapdeck_rcpp_triangle_interleaved`, data, params, list_columns, digits, layer_name)
 }
 
 rcpp_polygon_polyline <- function(data, params, geometry_columns) {