+++ R/map_layer_bitmap.R
@@ -99,9 +99,9 @@ add_bitmap <- function(
 
 #' @rdname clear
 #' @export
-clear_bitmap <- function( map, layer_id = NULL) {
+clear_bitmap <- function( map, layer_id = NULL, update_view = TRUE ) {
 	layer_id <- layerId(layer_id, "bitmap")
-	invoke_method(map, "md_layer_clear", layer_id, "bitmap" )
+	invoke_method(map, "md_layer_clear", layer_id, "bitmap", update_view )
 }