+++ R/map_layer_hexagon.R
@@ -238,7 +238,7 @@ add_hexagon <- function(
 
 #' @rdname clear
 #' @export
-clear_hexagon <- function( map, layer_id = NULL) {
+clear_hexagon <- function( map, layer_id = NULL, update_view = TRUE ) {
 	layer_id <- layerId(layer_id, "hexagon")
-	invoke_method(map, "md_layer_clear", map_type( map ), layer_id, "hexagon" )
+	invoke_method(map, "md_layer_clear", map_type( map ), layer_id, "hexagon", update_view )
 }