+++ inst/include/mapdeck_defaults.hpp
@@ -4,8 +4,26 @@
 #include <Rcpp.h>
 
 namespace mapdeck {
-namespace defaults {
+namespace binary_columns {
 
+	inline Rcpp::StringVector get_binary_columns( std::string layer_name ) {
+		if( layer_name == "trips" ) {
+			return Rcpp::StringVector{"stroke_width"};
+		} else if ( layer_name == "path" ) {
+		  return Rcpp::StringVector{"stroke_width","dash_size","dash_gap","offset"};
+		} else if (layer_name == "triangle" ) {
+			return Rcpp::StringVector{"elevation","fill_colour","stroke_colour"};
+		}
+
+		Rcpp::stop("mapdeck - unknown binary layer type");
+	}
+
+} // binary_columns
+} // mapdeck
+
+
+namespace mapdeck {
+namespace defaults {
 
 	inline Rcpp::NumericVector default_arc_height( int n ) {
 		Rcpp::NumericVector nv(n, 1.0);
@@ -52,15 +70,25 @@ namespace defaults {
 		return nv;
 	}
 
-	inline Rcpp::NumericVector default_fill_opacity( int n ) {
-		Rcpp::NumericVector nv(n, 255.0);
+	// inline Rcpp::NumericVector default_fill_opacity( int n ) {
+	// 	Rcpp::NumericVector nv(n, 255.0);
+	// 	return nv;
+	// }
+
+	// inline Rcpp::NumericVector default_stroke_opacity( int n ) {
+	// 	Rcpp::NumericVector nv(n, 255.0);
+	// 	return nv;
+	// }
+
+	inline Rcpp::NumericVector default_binary_colour( int n ) {
+		Rcpp::NumericVector nv(n, 1.0);
 		return nv;
 	}
 
-	inline Rcpp::NumericVector default_stroke_opacity( int n ) {
-		Rcpp::NumericVector nv(n, 255.0);
-		return nv;
-	}
+	// inline Rcpp::NumericVector default_binary_opacity( int n ) {
+	// 	Rcpp::NumericVector nv(n, 1.0);
+	// 	return nv;
+	// }
 
 	inline Rcpp::NumericVector default_stroke_width( int n ) {
 		Rcpp::NumericVector nv(n, 1.0);