+++ src/RcppExports.cpp
@@ -160,6 +160,37 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// rcpp_path_interleaved
+SEXP rcpp_path_interleaved(Rcpp::DataFrame sf, Rcpp::List params, Rcpp::IntegerVector list_columns, int digits, std::string layer_name);
+RcppExport SEXP _mapdeck_rcpp_path_interleaved(SEXP sfSEXP, SEXP paramsSEXP, SEXP list_columnsSEXP, SEXP digitsSEXP, SEXP layer_nameSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< Rcpp::DataFrame >::type sf(sfSEXP);
+    Rcpp::traits::input_parameter< Rcpp::List >::type params(paramsSEXP);
+    Rcpp::traits::input_parameter< Rcpp::IntegerVector >::type list_columns(list_columnsSEXP);
+    Rcpp::traits::input_parameter< int >::type digits(digitsSEXP);
+    Rcpp::traits::input_parameter< std::string >::type layer_name(layer_nameSEXP);
+    rcpp_result_gen = Rcpp::wrap(rcpp_path_interleaved(sf, params, list_columns, digits, layer_name));
+    return rcpp_result_gen;
+END_RCPP
+}
+// rcpp_trips_interleaved
+SEXP rcpp_trips_interleaved(Rcpp::DataFrame sf, Rcpp::List params, Rcpp::IntegerVector list_columns, int digits, std::string layer_name, int start_time);
+RcppExport SEXP _mapdeck_rcpp_trips_interleaved(SEXP sfSEXP, SEXP paramsSEXP, SEXP list_columnsSEXP, SEXP digitsSEXP, SEXP layer_nameSEXP, SEXP start_timeSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< Rcpp::DataFrame >::type sf(sfSEXP);
+    Rcpp::traits::input_parameter< Rcpp::List >::type params(paramsSEXP);
+    Rcpp::traits::input_parameter< Rcpp::IntegerVector >::type list_columns(list_columnsSEXP);
+    Rcpp::traits::input_parameter< int >::type digits(digitsSEXP);
+    Rcpp::traits::input_parameter< std::string >::type layer_name(layer_nameSEXP);
+    Rcpp::traits::input_parameter< int >::type start_time(start_timeSEXP);
+    rcpp_result_gen = Rcpp::wrap(rcpp_trips_interleaved(sf, params, list_columns, digits, layer_name, start_time));
+    return rcpp_result_gen;
+END_RCPP
+}
 // rcpp_path_polyline
 Rcpp::List rcpp_path_polyline(Rcpp::DataFrame data, Rcpp::List params, Rcpp::StringVector geometry_columns, std::string layer_name);
 RcppExport SEXP _mapdeck_rcpp_path_polyline(SEXP dataSEXP, SEXP paramsSEXP, SEXP geometry_columnsSEXP, SEXP layer_nameSEXP) {
@@ -247,18 +278,18 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
-// rcpp_triangle_columnar
-Rcpp::List rcpp_triangle_columnar(Rcpp::DataFrame data, Rcpp::List params, Rcpp::IntegerVector list_columns, Rcpp::StringVector geometry_columns, int digits);
-RcppExport SEXP _mapdeck_rcpp_triangle_columnar(SEXP dataSEXP, SEXP paramsSEXP, SEXP list_columnsSEXP, SEXP geometry_columnsSEXP, SEXP digitsSEXP) {
+// rcpp_triangle_interleaved
+Rcpp::List rcpp_triangle_interleaved(Rcpp::DataFrame data, Rcpp::List params, Rcpp::IntegerVector list_columns, int digits, std::string layer_name);
+RcppExport SEXP _mapdeck_rcpp_triangle_interleaved(SEXP dataSEXP, SEXP paramsSEXP, SEXP list_columnsSEXP, SEXP digitsSEXP, SEXP layer_nameSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< Rcpp::DataFrame >::type data(dataSEXP);
     Rcpp::traits::input_parameter< Rcpp::List >::type params(paramsSEXP);
     Rcpp::traits::input_parameter< Rcpp::IntegerVector >::type list_columns(list_columnsSEXP);
-    Rcpp::traits::input_parameter< Rcpp::StringVector >::type geometry_columns(geometry_columnsSEXP);
     Rcpp::traits::input_parameter< int >::type digits(digitsSEXP);
-    rcpp_result_gen = Rcpp::wrap(rcpp_triangle_columnar(data, params, list_columns, geometry_columns, digits));
+    Rcpp::traits::input_parameter< std::string >::type layer_name(layer_nameSEXP);
+    rcpp_result_gen = Rcpp::wrap(rcpp_triangle_interleaved(data, params, list_columns, digits, layer_name));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -329,13 +360,15 @@ static const R_CallMethodDef CallEntries[] = {
     {"_mapdeck_rcpp_od_geojson_df", (DL_FUNC) &_mapdeck_rcpp_od_geojson_df, 5},
     {"_mapdeck_rcpp_od_polyline", (DL_FUNC) &_mapdeck_rcpp_od_polyline, 4},
     {"_mapdeck_rcpp_path_geojson", (DL_FUNC) &_mapdeck_rcpp_path_geojson, 5},
+    {"_mapdeck_rcpp_path_interleaved", (DL_FUNC) &_mapdeck_rcpp_path_interleaved, 5},
+    {"_mapdeck_rcpp_trips_interleaved", (DL_FUNC) &_mapdeck_rcpp_trips_interleaved, 6},
     {"_mapdeck_rcpp_path_polyline", (DL_FUNC) &_mapdeck_rcpp_path_polyline, 4},
     {"_mapdeck_rcpp_point_df_columnar", (DL_FUNC) &_mapdeck_rcpp_point_df_columnar, 5},
     {"_mapdeck_rcpp_point_sf_columnar", (DL_FUNC) &_mapdeck_rcpp_point_sf_columnar, 5},
     {"_mapdeck_rcpp_point_geojson_df", (DL_FUNC) &_mapdeck_rcpp_point_geojson_df, 5},
     {"_mapdeck_rcpp_point_polyline", (DL_FUNC) &_mapdeck_rcpp_point_polyline, 4},
     {"_mapdeck_rcpp_polygon_geojson", (DL_FUNC) &_mapdeck_rcpp_polygon_geojson, 4},
-    {"_mapdeck_rcpp_triangle_columnar", (DL_FUNC) &_mapdeck_rcpp_triangle_columnar, 5},
+    {"_mapdeck_rcpp_triangle_interleaved", (DL_FUNC) &_mapdeck_rcpp_triangle_interleaved, 5},
     {"_mapdeck_rcpp_polygon_polyline", (DL_FUNC) &_mapdeck_rcpp_polygon_polyline, 3},
     {"_mapdeck_rcpp_text_geojson", (DL_FUNC) &_mapdeck_rcpp_text_geojson, 4},
     {"_mapdeck_rcpp_text_geojson_df", (DL_FUNC) &_mapdeck_rcpp_text_geojson_df, 4},